DDF-1058 Move wrapper back to its own module. This is necessary in order to avoid implicit dependency cycles when the wrapper is in security-commons.
